incomes facture generator

+++ backAPI/src/main/java/com/task/DTO/Records/ProductsRecord.java
@@ -4,5 +4,5 @@ public record ProductsRecord(
         String code,
         String name,
         String description,
-        Double price
+        Integer price
 ) {}
