+++ backAPI/src/main/java/com/task/DTO/Records/SalesRecord.java
@@ -3,7 +3,7 @@ package com.task.DTO.Records;
 public record SalesRecord(
         String code,
         String date,
-        Double price,
+        Integer price,
         String clientCode,
         String productCode
 ) {}
